Reorganize conversation list UI into subpackages
