+++ app/src/test/kotlin/com/android/messaging/ui/conversationlist/ui/item/ConversationListItemContentDescriptionTest.kt
@@ -1,7 +1,8 @@
-package com.android.messaging.ui.conversationlist.ui
+package com.android.messaging.ui.conversationlist.ui.item
 
 import com.android.messaging.R
 import com.android.messaging.data.conversationlist.model.ConversationListMessageStatus
+import com.android.messaging.ui.conversationlist.ui.support.previewConversationListItem
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertFalse
 import org.junit.Assert.assertTrue
